all priors to prior.py; all decoders to decoder.py
